[nick/#29] feat/package 분리
