+++ Nick/Nike/app/src/main/java/com/clone/nike/ui/cart/CartFragment.kt
@@ -1,12 +1,11 @@
-package com.clone.nike
+package com.clone.nike.ui.cart
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import com.clone.nike.databinding.FragmentCartBinding
+import com.clone.nike.R
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class CartFragment: Fragment() {
