+++ Nick/Nike/app/src/main/java/com/clone/nike/ui/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.clone.nike
+package com.clone.nike.ui.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
