+++ Nick/Nike/app/src/main/java/com/clone/nike/ui/purchase/PurchaseFragment.kt
@@ -1,11 +1,10 @@
-package com.clone.nike
+package com.clone.nike.ui.purchase
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.clone.nike.databinding.FragmentPurchaseBinding
 
 class PurchaseFragment: Fragment() {
     private lateinit var binding: FragmentPurchaseBinding
