+++ Nick/Nike/app/src/main/java/com/clone/nike/ui/purchase/WishFragment.kt
@@ -1,21 +1,20 @@
-package com.clone.nike
+package com.clone.nike.ui.purchase
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.clone.nike.databinding.FragmentProfileBinding
 
-class ProfileFragment: Fragment() {
-    private lateinit var binding: FragmentProfileBinding
+class WishFragment: Fragment() {
+    private lateinit var binding: FragmentWishBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentProfileBinding.inflate(layoutInflater)
+        binding = FragmentWishBinding.inflate(layoutInflater)
         return binding.root
     }
 }
