+++ Nick/Nike/app/src/main/res/layout/item_purchase_list.xml
@@ -62,6 +62,4 @@
             android:text="@string/item_purchase_price"
             android:layout_marginTop="5dp"/>
     </LinearLayout>
-    
-
 </androidx.constraintlayout.widget.ConstraintLayout>
